Tests: FIX: t6500-color-invalid blocks until Ctrl-D is pressed
Because the last test case is comprised of a successful reldate run that processes input.
In TAP::Harness v3.42 and Perl v5.30.0 this blocks, whereas it still passed with TAP::Harness v3.35 and Perl v5.22.1. By setting the stdin option to a reference to an empty string, that is taken as input.